split off the hexagon background into its own component and removed the code from HomePage.jsx to clean it up
