+++ frontend/src/pages/Home/HomePage.jsx
@@ -2,22 +2,10 @@ import { Link } from "react-router-dom";
 import { HivemindLogo } from "../../components/HivemindLogo";
 import { useState, useEffect } from "react";
 import "./HomePage.css";
-import "./Hexagons.css";
 import images from "../../images";
+import HexagonBackground from "../../components/HexagonBackground";
 
 export function HomePage() {
-  const hexagons = [];
-  let i = 0;
-  for(let x = -1; x < 14; x++) {
-    for(let y =-1; y < 16; y++) {
-      const hexagon = <div className="hexagon-filled" style={{left: x * 9.4 + 'rem', top: y * 5.5 + 'rem'}} key={i}></div>;
-      i++;
-      const hexagon2 = <div className="hexagon-filled" style={{left: x * 9.4 + 4.7 + 'rem', top: y * 5.5 + 2.75 + 'rem'}} key={i}></div>;
-      i++;
-      hexagons.push(hexagon);
-      hexagons.push(hexagon2);
-    };
-  };
 
   window.addEventListener('scroll', () => {
     document.getElementById("home-container").style.setProperty('--scroll', window.pageYOffset / (document.body.offsetHeight - window.innerHeight));
@@ -27,7 +15,7 @@ export function HomePage() {
   return (
     <div className="home-container" id="home-container">
       <div className="background-pattern"></div>
-      <div>{hexagons}</div>
+      <HexagonBackground />
       <div className="hive-icon">
         <div className="hive-bar-small brown"></div>
         <div className="hive-bar-large yellow"></div>
